+++ src/main/java/com/rodrigo/bookstore/repository/CategoriaRepository.java
@@ -5,7 +5,8 @@ import org.springframework.stereotype.Repository;
 
 import com.rodrigo.bookstore.domain.Categoria;
 
+//os Respository é a classe onde faz tudo do banco de Dados
 @Repository
-public interface CategoriaRepository extends JpaRepository<Categoria, Long> {
+public interface CategoriaRepository extends JpaRepository<Categoria, Integer> {
 
 }
